Update NufxLib snapshot to v3.0.0d2
This integrates the latest NufxLib sources, and updates CiderPress
to work with the API changes.

+++ app/AboutDialog.cpp
@@ -32,7 +32,7 @@ static const WCHAR kVersionExtra[] =
 BOOL AboutDialog::OnInitDialog(void)
 {
     NuError nerr;
-    long major, minor, bug;
+    int32_t major, minor, bug;
     CString newVersion, tmpStr;
     CStatic* pStatic;
     //CString versionFmt;
